+++ src/models/fieldHandlers/elements/Matrix.php
@@ -39,6 +39,9 @@ class Matrix extends ElementFieldHandler
      */
     public function isDirty(FieldHandler $handler): bool
     {
+        if (get_class($handler) != get_class($this)) {
+            return true;
+        }
         return !empty($this->getDirty($handler)['blocks']);
     }
 
@@ -88,11 +91,11 @@ class Matrix extends ElementFieldHandler
             ];
             foreach ($block['fields'] as $fieldId => $handler) {
                 $oldHandler = $oldBlocks[$id]['fields'][$fieldId] ?? null;
-                if ($oldHandler and $fdirty = $handler->getDirty($oldHandler)) {
+                if ($oldHandler and $handler->isDirty($oldHandler)) {
                     $blockIsdirty = true;
                     $blockDirty['fields'][$fieldId] = [
                         'handler' => get_class($handler),
-                        'data' => $fdirty
+                        'data' => $handler->getDirty($oldHandler)
                     ];
                 }
             }
@@ -148,18 +151,11 @@ class Matrix extends ElementFieldHandler
     protected function buildValues(): array
     {
         $value = [];
-        $blocks = $this->field->normalizeValue($this->element->getFieldValue($this->field->handle), $this->element)->all();
-        foreach ($blocks as $id => $block) {
+        $blocks = $this->field->normalizeValue($this->rawValue, $this->element)->all();
+        foreach ($blocks as $block) {
             $fields = [];
             foreach ($this->field->getBlockTypeFields([$block->type->id]) as $field) {
-                $class = Activity::$plugin->fieldHandlers->getForElementField(get_class($field));
-                $fields[$field->id] = new $class([
-                    'field' => $field,
-                    'element' => $block,
-                    'name' => $field->name,
-                    'value' => $field->serializeValue($block->{$field->handle}, $block),
-                    'rawValue' => $block->{$field->handle}
-                ]);
+                $fields[$field->id] = Activity::$plugin->fieldHandlers->getHandlerForField($field, $block);
             }
             $value[] = [
                 'enabled' => $block->enabled,